requirements.txt
asgiref==3.4.1
Django==3.2.6
django-redis==5.0.0
djangorestframework==3.12.4
psycopg2==2.9.1
pytz==2021.1
redis==3.5.3
sqlparse==0.4.1
typing-extensions==3.10.0.0
